Rename fetch utility and update imports
