+++ src/utils/fetchStudy.ts
